Moved entrypoint into separate project to enable Console vs. NT Service packaging
