+++ Dns/Dns.csproj
@@ -1,10 +1,9 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <OutputType>Exe</OutputType>
+    <OutputType>Library</OutputType>
     <TargetFramework>netcoreapp3.1</TargetFramework>
     <RootNamespace>Dns</RootNamespace>
-    <StartupObject>Dns.Program</StartupObject>
     <CodePage>65001</CodePage>
   </PropertyGroup>
