+++ dns-cli/appsettings.json
@@ -1,0 +1,42 @@
+﻿{
+  "server": {
+    "zone": {
+      "name": ".stephbu.org",
+      "provider": "Dns.ZoneProvider.IPProbe.IPProbeZoneProvider"
+    },
+    "dnslistener": {
+      "port": 5335
+    },
+    "webserver": {
+      "port": 8080,
+      "enabled":  true
+    }
+  },
+  "zoneprovider": {
+    "PollingIntervalSeconds": 15,
+    "Hosts": [
+      {
+        "Name": "foo",
+        "Probe": "ping",
+        "Timeout": 30,
+        "AvailabilityMode": "first",
+        "Ip": [
+          "192.168.1.1",
+          "192.168.1.252",
+          "192.168.1.253",
+          "192.168.86.100"
+        ]
+      },
+      {
+        "Name": "test",
+        "Probe": "http",
+        "Timeout": 100,
+        "AvailabilityMode": "all",
+        "Ip": [
+          "192.168.86.1",
+          "192.168.1.1"
+        ]
+      }
+    ]
+  }
+}
